+++ gluecodium/src/test/resources/smoke/comments/output/android/com/example/smoke/CommentsLinks.java
@@ -60,6 +60,8 @@ public final class CommentsLinks extends NativeBase {
      * <li>top level enum: {@link com.example.smoke.TypeCollectionEnum}</li>
      * <li>top level enum item: {@link com.example.smoke.TypeCollectionEnum#ITEM}</li>
      * <li>error: {@link com.example.smoke.Comments.SomethingWrongException}</li>
+     * <li>type from aux sources, same package: {@link com.example.smoke.AuxClass}</li>
+     * <li>type from aux sources, different package: {@link com.example.fire.AuxStruct}</li>
      * </ul>
      * <p>Not working for Java:</p>
      * <ul>