+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Gluecodium project Release Notes
 
 ## Unreleased
+### Bug fixes:
+  * Documentation links to elements declared in "auxiliary" IDL sources are now resolved correctly
+    in the generated C++, Java, and Swift code.
 ### Breaking changes:
   * Added validation against name clashes between property declarations in IDL files.